+++ plotlykt-core/src/commonMain/kotlin/space/kscience/plotly/Plotly.kt
@@ -13,6 +13,10 @@ import kotlin.js.JsName
 @JsName("PlotlyKt")
 public object Plotly {
     public const val VERSION: String = "1.54.6"
+
+    public const val PLOTLY_CDN: String = "https://cdn.plot.ly/plotly-${VERSION}.min.js"
+    //"https://cdnjs.cloudflare.com/ajax/libs/plotly.js/${VERSION}/plotly.min.js"
+
     public inline fun plot(block: Plot.() -> Unit): Plot = Plot().apply(block)
 }