+++ plotlykt-core/src/jvmMain/kotlin/space/kscience/plotly/PlotlyPage.kt
@@ -12,7 +12,7 @@ public class PlotlyFragment(public val render: FlowContent.(renderer: PlotlyRend
  * A complete page including headers and title
  */
 public data class PlotlyPage(
-    val headers: Collection<HtmlFragment>,
+    val headers: Collection<PlotlyHtmlFragment>,
     val fragment: PlotlyFragment,
     val title: String = "Plotly.kt",
     val renderer: PlotlyRenderer = StaticPlotlyRenderer
@@ -37,7 +37,7 @@ public fun Plotly.fragment(content: FlowContent.(renderer: PlotlyRenderer) -> Un
  * Create a complete page including plots
  */
 public fun Plotly.page(
-    vararg headers: HtmlFragment = arrayOf(cdnPlotlyHeader),
+    vararg headers: PlotlyHtmlFragment = arrayOf(cdnPlotlyHeader),
     title: String = "Plotly.kt",
     renderer: PlotlyRenderer = StaticPlotlyRenderer,
     content: FlowContent.(renderer: PlotlyRenderer) -> Unit
@@ -47,7 +47,7 @@ public fun Plotly.page(
  * Convert an html plot fragment to page
  */
 public fun PlotlyFragment.toPage(
-    vararg headers: HtmlFragment = arrayOf(cdnPlotlyHeader),
+    vararg headers: PlotlyHtmlFragment = arrayOf(cdnPlotlyHeader),
     title: String = "Plotly.kt",
     renderer: PlotlyRenderer = StaticPlotlyRenderer
 ): PlotlyPage = PlotlyPage(headers.toList(), this, title, renderer)
@@ -56,7 +56,7 @@ public fun PlotlyFragment.toPage(
  * Convert a plot to the sigle-plot page
  */
 public fun Plot.toPage(
-    vararg headers: HtmlFragment = arrayOf(cdnPlotlyHeader),
+    vararg headers: PlotlyHtmlFragment = arrayOf(cdnPlotlyHeader),
     config: PlotlyConfig = PlotlyConfig.empty(),
     title: String = "Plotly.kt",
     renderer: PlotlyRenderer = StaticPlotlyRenderer