+++ plotlykt-core/src/jvmMain/kotlin/space/kscience/plotly/bootstrapHeaders.kt
@@ -31,7 +31,7 @@ import kotlinx.html.script
 //    }
 //}
 
-public val cdnBootstrap: HtmlFragment = HtmlFragment {
+public val cdnBootstrap: PlotlyHtmlFragment = PlotlyHtmlFragment {
     script {
         src = "https://code.jquery.com/jquery-3.5.1.slim.min.js"
         integrity = "sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj"