+++ plotlykt-core/src/jvmMain/resources/js/plotlyConnect.js
@@ -162,4 +162,3 @@ function startPush(id, ws) {
     };
 }
 
-