+++ plotlykt-jupyter/src/main/kotlin/space/kscience/plotly/PlotlyIntegration.kt
@@ -2,9 +2,15 @@ package space.kscience.plotly
 
 import kotlinx.html.*
 import kotlinx.html.stream.createHTML
+import org.jetbrains.kotlinx.jupyter.api.HTML
+import org.jetbrains.kotlinx.jupyter.api.annotations.JupyterLibrary
+import org.jetbrains.kotlinx.jupyter.api.libraries.*
+import space.kscience.plotly.Plotly.PLOTLY_CDN
 
 @UnstablePlotlyAPI
-public object JupyterPlotly: PlotlyRenderer {
+@JupyterLibrary
+internal class PlotlyIntegration : JupyterIntegration(), PlotlyRenderer {
+
     override fun FlowContent.renderPlot(plot: Plot, plotId: String, config: PlotlyConfig): Plot {
         div {
             id = plotId
@@ -38,7 +44,7 @@ public object JupyterPlotly: PlotlyRenderer {
         return plot
     }
 
-    public fun loadJs(): HtmlFragment = HtmlFragment {
+    private fun loadJs(): PlotlyHtmlFragment = PlotlyHtmlFragment {
         script {
             type = "text/javascript"
             unsafe {
@@ -83,17 +89,79 @@ public object JupyterPlotly: PlotlyRenderer {
         }
     }
 
-    public fun renderPlot(plot: Plot): String = createHTML().div {
-        plot(plot, config = PlotlyConfig{
+    private fun renderPlot(plot: Plot): String = createHTML().div {
+        plot(plot, config = PlotlyConfig {
             responsive = true
-        }, renderer = this@JupyterPlotly)
+        }, renderer = this@PlotlyIntegration)
     }
 
-    public fun renderFragment(fragment: PlotlyFragment): String = createHTML().div {
+    private fun renderFragment(fragment: PlotlyFragment): String = createHTML().div {
         with(fragment) {
-            render(this@JupyterPlotly)
+            render(this@PlotlyIntegration)
         }
     }
 
-    public fun renderPage(page: PlotlyPage): String = page.copy(renderer = this@JupyterPlotly).render()
+    private fun renderPage(page: PlotlyPage): String = page.copy(renderer = this@PlotlyIntegration).render()
+
+
+    private val plotlyBundle = ResourceFallbacksBundle(listOf(
+        ResourceLocation(
+            PLOTLY_CDN,
+            ResourcePathType.URL
+        ),
+        ResourceLocation(
+            "js/plotly.min.js",
+            ResourcePathType.CLASSPATH_PATH
+        )
+    ))
+
+    private val plotlyResource =
+        LibraryResource(name = "plotly", type = ResourceType.JS, bundles = listOf(plotlyBundle))
+
+    private val plotlyConnectBundle = ResourceFallbacksBundle(listOf(
+        ResourceLocation(
+            "js/plotlyConnect.js",
+            ResourcePathType.CLASSPATH_PATH
+        )
+    ))
+
+    private val plotlyConnectResource =
+        LibraryResource(name = "plotlyConnect", type = ResourceType.JS, bundles = listOf(plotlyConnectBundle))
+
+
+    override fun Builder.onLoaded() {
+
+        resource(plotlyResource)
+        resource(plotlyConnectResource)
+
+        repositories("https://repo.kotlin.link")
+
+        import(
+            "space.kscience.plotly.*",
+            "space.kscience.plotly.models.*",
+            "space.kscience.dataforge.meta.*",
+            "kotlinx.html.*"
+        )
+
+        onLoaded {
+            display(HTML(loadJs().toString()))
+        }
+
+        render<PlotlyHtmlFragment> {
+            HTML(it.toString())
+        }
+
+        render<Plot> {
+            HTML(renderPlot(it))
+        }
+
+        render<PlotlyFragment> {
+            HTML(renderFragment(it))
+        }
+
+        render<PlotlyPage> {
+            HTML(renderPage(it), true)
+        }
+    }
+
 }